Fix BufferInvalidationQueue partitioning algorithm.
Description: After coalescing a node,
the next node is deleted from the
queue. However, in order to prevent
indices shifting in the "array" of nodes,
the reference to the next node (which
is the indexOfTheNodeCoalescing + 1)
is assigned to the reference of the
"merged" node. However, this failed
to capture the recursive nature of
reassigning the references. If multiple
references already point to the "next"
node, then all of them must be set. This
bug addressing this mistake.

Other changes: The condition to check
if the node is "last" in the queue was
removed. That is because it is
impossible for it to be "last" in the queue -
and that is ensured by fact that the
last node is not added to the array
of nodes; moreover, the same index
is never merged twice. This guarantees
an untouched index will never be the
last node in the queue. Note that this "check"
was added in the first place b/c some
nodes were being referenced in the
node array, that were removed from
the queue (as described in the description)